core/local: Rename ./steps to ./atom
So we can group everything atom-related here.
It was containing other things than steps anyway.
