+++ core/local/atom/add_checksum.js
