+++ core/local/atom/add_infos.js
