+++ core/local/atom/await_write_finish.js
