+++ core/local/atom/event.js
