+++ core/local/atom/initial_diff.js
