+++ core/local/atom/producer.js
