+++ core/local/atom/scan_folder.js
