+++ core/local/atom/win_detect_move.js
