+++ core/local/atom/win_identical_renaming.js
