+++ core/local/watcher.js
@@ -5,7 +5,7 @@ const ChokidarWatcher = require('./chokidar_watcher')
 
 /*::
 import type { WatcherType } from '../config'
-import type { AtomEventsDispatcher } from './steps/dispatch'
+import type { AtomEventsDispatcher } from './atom/dispatch'
 import type Pouch from '../pouch'
 import type Prep from '../prep'
 import type EventEmitter from 'events'
